don't version private endpoints
