+++ src/middlewared/middlewared/api/v25_10_0/boot.py
@@ -1,30 +1,17 @@
-from typing import Literal
-
 from pydantic import Field, PositiveInt
 
-from middlewared.api.base import BaseModel, NotRequired
+from middlewared.api.base import BaseModel
 
 
 __all__ = [
     "BootGetDisksArgs", "BootGetDisksResult", "BootAttachArgs", "BootAttachResult", "BootDetachArgs",
     "BootDetachResult", "BootReplaceArgs", "BootReplaceResult", "BootScrubArgs", "BootScrubResult",
-    "BootSetScrubIntervalArgs", "BootSetScrubIntervalResult", "BootUpdateInitramfsArgs", "BootUpdateInitramfsResult",
-    "BootFormatArgs", "BootFormatResult"
+    "BootSetScrubIntervalArgs", "BootSetScrubIntervalResult",
 ]
 
 
 class BootAttachOptions(BaseModel):
     expand: bool = False
-
-
-class BootFormatOptions(BaseModel):
-    size: int = NotRequired
-    legacy_schema: Literal["BIOS_ONLY", "EFI_ONLY", None] = None
-
-
-class BootUpdateInitramfsOptions(BaseModel):
-    database: str | None = None
-    force: bool = False
 
 
 class BootGetDisksArgs(BaseModel):
@@ -52,15 +39,6 @@ class BootDetachResult(BaseModel):
     result: None
 
 
-class BootFormatArgs(BaseModel):
-    dev: str
-    options: BootFormatOptions = Field(default_factory=BootFormatOptions)
-
-
-class BootFormatResult(BaseModel):
-    result: None
-
-
 class BootReplaceArgs(BaseModel):
     label: str
     dev: str
@@ -84,11 +62,3 @@ class BootSetScrubIntervalArgs(BaseModel):
 
 class BootSetScrubIntervalResult(BaseModel):
     result: PositiveInt
-
-
-class BootUpdateInitramfsArgs(BaseModel):
-    options: BootUpdateInitramfsOptions = Field(default_factory=BootUpdateInitramfsOptions)
-
-
-class BootUpdateInitramfsResult(BaseModel):
-    result: bool
